Insert code in an R Markdown document to diagnose personal data information

This function is used, inside an R Markdown document, to dynamically elaborate a short report on the client-side (by using JavaScript code) that displays the login status of the user (login, contextual data, ...).

diagnose_login(block = "info", lang = "en")

Arguments

block

Should the diagnostic be included in a block construct (not if NULL)?

lang

The language to use for diagnostic output (only 'en' or 'fr' currently)

Value

Nothing, the function is used for its side-effect of adding diagnostic code in the document

Details

Wherever you want to display this diagnostic report in your bookdown page, add on its own line the following instruction: 'r learnitdown::diagnose_login()' between backquotes (you can also change lang, if you want).