Set up a learnitdown Learnr application

This function eases the configuration of the learnr document to get user and database info, record events, use grade this and parameterize learnr.

learnitdownLearnrSetup(
  config,
  sign_in,
  time.limit = 60,
  cap = "R Code",
  echo = FALSE,
  comment = NA,
  use.gradethis = TRUE,
  event.recorder = learnitdown::record_learnr,
  debug = Sys.getenv("LEARNITDOWN_DEBUG", 0) != 0
)

learnitdownLearnrBanner(
  title,
  text,
  image,
  align = "left",
  msg.nologin = "Anonymous user, no record!",
  msg.login = "Recording activated for ",
  msg.error = "Error recording activity! "
)

learnitdownLearnrServer(input, output, session)

Arguments

config

The config() command to use to get database info.

sign_in

The sign_in() command to use to get user info.

time.limit

The maximum time allowed to evaluate R code.

cap

The caption for R code widgets.

echo

Do we echo commands in R chunks?

comment

The prefix added before each line of R chunk output.

use.gradethis

Do we use {gradethis}?

event.recorder

The function to use as event recorder. you should probably not change the default value here.

debug

Do we issue additions debugging information?

title

The Title for the banner.

text

Text to print beneath the title.

image

URL to an image to display in the banner.

align

How is the image aligned: "left" (default), "right", "middle", "top" or "bottom".

msg.nologin

The message to display if no user is logged in.

msg.login

The message to display if a user is logged in (will be followed by the login).

msg.error

The message when an error during recording of activity in the database occurs.

input

The Shiny input.

output

The Shiny output.

session

The Shiny session.

Value

Nothing. The function is used to setup the learnr environment.