Insert a block for a learnr tutorial

Insert Markdown text to link to a learnr tutorial.

learnr(
  id,
  title = NULL,
  package,
  toc = "",
  text = "Now let's make the exercises in the following tutorial:",
  toc.def = "Tutorial {id}",
  rstudio.url = "start_rstudio.html",
  tuto.img = "images/list-tuto.png",
  tuto.link = "tutorial"
)

Arguments

id

Identifier of the learnr tutorial (as tutorial:id in the YAML section).

title

The title of the tutorial (title in the YAML section). If NULL, the id is used instead.

package

Package where the learnr tutorial is defined.

toc

Entry to use in the exercises table of content (NULL if no entry, "" for a default entry based on toc.def =).

text

The text to display in the learnr block.

toc.def

Text for a default toc entry using glue() syntax for replacement, e.g., {id}.

rstudio.url

The URL to open a page in RStudio server in the SciViews box.

tuto.img

The image to display in front of the toc entry

tuto.link

The link when the image is clicked (sends to an help page about learnr tutorials).

Value

The Markdown chunk to insert a learnr tutorial block in the document.