Record results of learnr exercises in a MongoDB database

Record tutorial submissions in a MongoDB database. The function is used by learnitdown learnr tutorials and is not for end-users.

record_learnr(tutorial_id, tutorial_version, user_id, event, data)

user_name(value)

user_email(value)

Arguments

tutorial_id

The identifier of the tutorial.

tutorial_version

The version of the tutorial.

user_id

The user identifier for this learnr process.

event

The event that triggers the record, like exercise_submission or question_submission

data

A JSON field with event-dependent data content. If NULL, only a test to see if the database is responding is performed.

value

The new value for user name or email (if not provided, the current value is returned).

Value

Nothing. The function is used for its side-effects.

See also