Record Shiny events in a MongoDB database

Given a path that contains shinylogs events in .rds format, read these events and transfer them into a MongoDB database.

record_shiny(
  path,
  url,
  db,
  collection = "events",
  version = "0",
  log.errors = TRUE,
  log.outputs = FALSE,
  drop.dir = FALSE,
  debug = Sys.getenv("LEARNITDOWN_DEBUG", 0) != 0
)

Arguments

path

The directory that contains shinylogs .rds files

url

The mongodb url.

db

The database name.

collection

The name of the collection where to insert the documents.

version

The version of the running Shiny application.

log.errors

Do we record errors too?

log.outputs

Do we record outputs too (note that results are recorded with inputs)?

drop.dir

If TRUE and path is empty at the end of the process, drop the logs directory.

debug

Do we debug the events recording by issuing extra messages? By Default the value in the environment variable LEARNITDOWN_DEBUG is used and debugging will be done if this value is different to 0.

Value

TRUEif there where log files to export, FALSE otherwise.

See also